Moved querysets into own module in jobs and clusters apps.
